[Fix] #13 - 브랜치 충돌 해결 (임시)
