+++ app/profile/types.ts
@@ -59,6 +59,10 @@ export interface ProfileState {
   settlementData: MySettlementResponse | null
   settlementMonth: SettlementMonth | null
 
+  // Farm state
+  hasFarm: boolean | null
+  isCheckingFarm: boolean
+
   // UI state
   activeTab: string
   isSellerDialogOpen: boolean
